% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{check_balance}
\alias{check_balance}
\title{Vérifier l'équilibre des 0 et 1}
\usage{
check_balance(mat, n)
}
\arguments{
\item{mat}{Une matrice de Takuzu}

\item{n}{La taille de la grille (n x n)}
}
\value{
TRUE si la règle est respectée, FALSE sinon
}
\description{
Vérifier l'équilibre des 0 et 1
}
