% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{check_consecutive}
\alias{check_consecutive}
\title{Vérifier la règle des trois consécutifs}
\usage{
check_consecutive(mat, n)
}
\arguments{
\item{mat}{Une matrice de Takuzu}

\item{n}{La taille de la grille (n x n)}
}
\value{
TRUE si la règle est respectée, FALSE sinon
}
\description{
Vérifier la règle des trois consécutifs
}
