% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{check_no_repeated}
\alias{check_no_repeated}
\title{vérifier si les lignes et les colonne sont identiques}
\usage{
check_no_repeated(mat, n)
}
\arguments{
\item{mat}{Une matrice de Takuzu}

\item{n}{La taille de la grille (n x n)}
}
\value{
TRUE si la règle est respectée, FALSE sinon
}
\description{
vérifier si les lignes et les colonne sont identiques
}
